requirements.txt
elasticsearch==7.17.9
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
pytz==2024.1